third_party/blink/web_tests/wpt_internal/prerender/restriction-async-clipboard.https.html

<!DOCTYPE html>
<!-- This cannot be upstreamed as a WPT test because the spec (probably) will
     require that no error is thrown until activation. -->
<title>Same-origin prerendering cannot access the Async Clipboard API</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to access the clipboard via the
  // Async Clipboard API.
  startPrerendering(`resources/async-clipboard.html`);
  const result = await gotMessage;
  assert_equals(result, 'NotAllowedError');
}, `prerendering pages should not be able to access the clipboard via the ` +
   `Async Clipboard API.`);

</script>
</body>
